added(conversion.h) for string-integer conversion added(ansi_color_codes.h) for color coded output to ANSI terminals added(break.h) to allow for programmatic breakpoints added(unreachable.h) for a macro version of unreachable that panics in debug mode added(test.h) for a test assert macro in simple test rigs. added(result_to_string_view) to allow the printing of result codes implemented(stream.c) for Linux systems implemented(message.c) implemented(panic.c) implemented(conversion.c) implemented(memory_tests.c)

+++ include/support/types/scalar.h
@@ -95,6 +95,6 @@ typedef double f64;
 #define i8_MIN  0x80
 #define i16_MIN 0x8000
 #define i32_MIN 0x80000000
-#define i64_MIN 0x8000000000000000ll
+#define i64_MIN (i64)(0x8000000000000000ll)
 
 #endif // !SUPPORT_TYPES_SCALAR_H
